refactor: move veil handling to the popup components
