+++ js/ui/export-popup.js
@@ -42,6 +42,29 @@ export const ExportPopup = {
         }
       }
   },
+  mounted () {
+    const popupEle = document.querySelector('.export_popup_div')
+    if (popupEle) {
+      const disablePopupFunc = function () {
+        globalPopup.export = false
+        window.MetricQWebView.instances[0].reload()
+      }
+      veil.create(disablePopupFunc)
+      veil.attachPopup(popupEle)
+      const closeButtonEle = popupEle.querySelector('.popup_close_button')
+      closeButtonEle.addEventListener('click', function () {
+        veil.destroy()
+        disablePopupFunc()
+      })
+      const modalEle = document.querySelector('.modal')
+      modalEle.addEventListener('click', function (evt) {
+        if (evt.target.getAttribute('role') === 'dialog') {
+          veil.destroy()
+          disablePopupFunc()
+        }
+      })
+    }
+  },
   methods: {
     doExport: function () {
       const instance = window.MetricQWebView.instances[0]
