+++ js/ui/xaxis-popup.js
@@ -1,4 +1,6 @@
 import { PopupHeader } from './popup-header.js'
+import { veil } from './veil.js'
+import { globalPopup } from '../app.js'
 
 // @vue/component
 export const XaxisPopup = {
@@ -52,6 +54,32 @@ export const XaxisPopup = {
         window.MetricQWebView.instances[0].handler.setTimeRange(undefined, dateObj.getTime())
         window.MetricQWebView.instances[0].setPlotRanges(true, true)
       }
+    }
+  },
+  mounted () {
+    const popupEle = document.querySelector('.xaxis_popup_div')
+    if (popupEle) {
+      const disablePopupFunc = function () {
+        globalPopup.xaxis = false
+        window.MetricQWebView.instances[0].reload()
+      }
+      veil.create(disablePopupFunc)
+      veil.attachPopup(popupEle)
+      const closeButtonEle = popupEle.querySelector('.popup_close_button')
+      const okEle = popupEle.querySelector('.popup_ok');
+      [closeButtonEle, okEle].forEach(function (paramValue, paramIndex, paramArr) {
+        paramValue.addEventListener('click', function () {
+          veil.destroy()
+          disablePopupFunc()
+        })
+      })
+      const modalEle = document.querySelector('.modal')
+      modalEle.addEventListener('click', function (evt) {
+        if (evt.target.getAttribute('role') === 'dialog') {
+          veil.destroy()
+          disablePopupFunc()
+        }
+      })
     }
   },
   template: `<div class="modal popup_div xaxis_popup_div" tabindex="-1" role="dialog">
