+++ js/ui/yaxis-popup.js
@@ -1,4 +1,6 @@
 import { PopupHeader } from './popup-header.js'
+import { veil } from './veil.js'
+import { globalPopup } from '../app.js'
 
 // @vue/component
 export const YaxisPopup = {
@@ -76,6 +78,32 @@ export const YaxisPopup = {
         window.MetricQWebView.instances[0].graticule.setYRangeOverride(undefined, arr[0], arr[1])
         window.MetricQWebView.instances[0].setPlotRanges(false, true)
       }
+    }
+  },
+  mounted () {
+    const popupEle = document.querySelector('.yaxis_popup_div')
+    if (popupEle) {
+      const disablePopupFunc = function () {
+        globalPopup.yaxis = false
+        window.MetricQWebView.instances[0].reload()
+      }
+      veil.create(disablePopupFunc)
+      veil.attachPopup(popupEle)
+      const closeButtonEle = popupEle.querySelector('.popup_close_button')
+      const okEle = popupEle.querySelector('.popup_ok');
+      [closeButtonEle, okEle].forEach(function (paramValue, paramIndex, paramArr) {
+        paramValue.addEventListener('click', function () {
+          veil.destroy()
+          disablePopupFunc()
+        })
+      })
+      const modalEle = document.querySelector('.modal')
+      modalEle.addEventListener('click', function (evt) {
+        if (evt.target.getAttribute('role') === 'dialog') {
+          veil.destroy()
+          disablePopupFunc()
+        }
+      })
     }
   },
   /* use vue-js for radio buttons */
